\name{RPathVisio-Server package}
\alias{RPathVisio-Server package}
\alias{RPathVisio-Server}
\docType{package}
\title{
R package for setting up a PathVisio XML-RPC server
}
\description{ 
With this package you can set up a PathVisio XML-RPC server from within R.
}
\author{
Anwesha Dutta <anwesha.dutta@maastrichtuniversity.nl> \\
Egon Willighagen <egon.willighagen@gmail.com> \\
Christ Leemans <C.Leemans@student.maastrichtuniversity.nl>
}
\keyword{ package }
\keyword{ PathVisio }
\keyword{ XML-RPC }
